lisp libraries refactored to favour embedded Lisp
