+++ cmd/lisp/main.go
@@ -6,10 +6,10 @@ import (
 
 	"github.com/jig/lisp/command"
 	"github.com/jig/lisp/env"
+	"github.com/jig/lisp/lib/assert/nsassert"
 	"github.com/jig/lisp/lib/concurrent/nsconcurrent"
 	"github.com/jig/lisp/lib/core/nscore"
 	"github.com/jig/lisp/lib/coreextented/nscoreextended"
-	"github.com/jig/lisp/lib/test/nstest"
 	"github.com/jig/lisp/types"
 )
 
@@ -25,7 +25,7 @@ func main() {
 		{"command line args", nscore.LoadCmdLineArgs},
 		{"concurrent", nsconcurrent.Load},
 		{"core mal extended", nscoreextended.Load},
-		{"test", nstest.Load},
+		{"assert", nsassert.Load},
 	} {
 		if err := library.load(ns); err != nil {
 			log.Fatalf("Library Load Error: %v\n", err)
