+++ lib/assert/nsassert/nsassert.go
@@ -1,11 +1,11 @@
-package nsconcurrent
+package nsassert
 
 import (
 	"context"
 	"reflect"
 
 	"github.com/jig/lisp"
-	"github.com/jig/lisp/lib/concurrent"
+	assert "github.com/jig/lisp/lib/assert"
 	"github.com/jig/lisp/types"
 )
 
@@ -14,11 +14,8 @@ type Here struct{}
 var _package_ = reflect.TypeOf(Here{}).PkgPath()
 
 func Load(env types.EnvType) error {
-	concurrent.Load(env)
-
-	if _, err := lisp.REPL(context.Background(), env, concurrent.HeaderConcurrent(), types.NewCursorFile(_package_)); err != nil {
+	if _, err := lisp.REPL(context.Background(), env, assert.HeaderAssertMacros(), types.NewCursorFile(_package_)); err != nil {
 		return err
 	}
-
 	return nil
 }
