+++ lib/concurrent/header-concurrent.lisp
@@ -1,0 +1,1 @@
+(defmacro future (fn [& body] `(^{:once true} future-call (fn [] ~@body))))
