+++ lib/core/header-load-file.lisp
@@ -1,0 +1,4 @@
+(def load-file (fn (f)
+                        (eval
+                            (read-string
+                                (str ";; $MODULE " f "\n(do " (slurp f) " nil)")))))
