--- lib/test/nstest/nstest.go
@@ -1,64 +1,0 @@
-package nstest
-
-import (
-	"context"
-	"reflect"
-
-	"github.com/jig/lisp"
-	"github.com/jig/lisp/types"
-	. "github.com/jig/lisp/types"
-)
-
-type Here struct{}
-
-var _package_ = reflect.TypeOf(Here{}).PkgPath()
-
-func Load(repl_env EnvType) error {
-	ctx := context.Background()
-	if _, err := lisp.REPL(ctx, repl_env, `(eval (read-string (str "(do "`+assertMacros+`" nil)")))`, types.NewCursorFile(_package_)); err != nil {
-		return err
-	}
-	return nil
-}
-
-var assertMacros = `;; assert macros
-(defmacro assert-true
-    (fn [name expr]
-        (list
-            'if (try expr (catch err err))
-                nil
-                {   :failed true
-                    :name name
-                    :expr (str expr)})))
-
-(defmacro assert-false
-    (fn [name expr]
-        (list
-            'if (try expr (catch err err))
-                {   :failed true
-                    :name name
-                    :expr (str expr)}
-                nil)))
-
-(defmacro assert-throws
-    (fn [name expr]
-        (let [failureError {   :failed true
-                                :name (str name)
-                                :expr (str expr)}]
-        ` + "`" + `(try
-            (do
-                ~expr
-                ~failureError)
-            (catch err nil)))))
-
-(def test.suite (fn [name & assert-cases]
-    (if
-        (reduce and true
-            (map
-                (fn [x]
-                    (if  (not (nil? x))
-                        (println "TEST SUITE FAIL" name ">" (get x :name) ">>" (get x :expr))
-                        true))
-                assert-cases))
-        (println "TEST SUITE PASS" name "PASS"))))
-`
