+++ macro_test.go
@@ -552,9 +552,6 @@ func TestMacro(t *testing.T) {
 	if _, err := REPL(ctx, repl_env, "(defmacro cond (fn (& xs) (if (> (count xs) 0) (list 'if (first xs) (if (> (count xs) 1) (nth xs 1) (throw \"odd number of forms to cond\")) (cons 'cond (rest (rest xs)))))))", types.NewCursorFile(reflect.TypeOf(_here_{}).PkgPath())); err != nil {
 		t.Fatal(err)
 	}
-	if _, err := REPL(ctx, repl_env, `(def db (atom {}))`, types.NewCursorFile(reflect.TypeOf(_here_{}).PkgPath())); err != nil {
-		t.Fatal(err)
-	}
 	if err := Load(repl_env); err != nil {
 		t.Fatal(err)
 	}
